Batch params
