+++ tests/transforms/test_batch_params.py
@@ -45,6 +45,30 @@ def test_simple_circuit(mocker):
     assert len(spy.call_args[0][0]) == batch_size
 
 
+def test_simple_circuit_one_batch(mocker):
+    """Test that batching works for a simple circuit when the batch size is 1"""
+    dev = qml.device("default.qubit", wires=3)
+
+    @qml.batch_params
+    @qml.qnode(dev, interface="autograd")
+    def circuit(data, x, weights):
+        qml.templates.AmplitudeEmbedding(data, wires=[0, 1, 2], normalize=True)
+        qml.RX(x, wires=0)
+        qml.RY(0.2, wires=1)
+        qml.templates.StronglyEntanglingLayers(weights, wires=[0, 1, 2])
+        return qml.probs(wires=[0, 2])
+
+    batch_size = 1
+    data = np.random.random((batch_size, 8))
+    x = np.linspace(0.1, 0.5, batch_size, requires_grad=True)
+    weights = np.ones((batch_size, 10, 3, 3), requires_grad=True)
+
+    spy = mocker.spy(circuit.device, "batch_execute")
+    res = circuit(data, x, weights)
+    assert res.shape == (batch_size, 4)
+    assert len(spy.call_args[0][0]) == batch_size
+
+
 def test_basic_entangler_layers(mocker):
     """Test that batching works for BasicEngtanglerLayers"""
     dev = qml.device("default.qubit", wires=2)
@@ -156,181 +180,545 @@ def test_basis_state_preparation(mocker):
     assert np.allclose(res, indiv_res)
 
 
-@pytest.mark.autograd
-@pytest.mark.parametrize("diff_method", ["backprop", "adjoint", "parameter-shift"])
-def test_autograd(diff_method, tol):
-    """Test derivatives when using autograd"""
-    dev = qml.device("default.qubit", wires=2)
+def test_multi_returns():
+    """Test that batching works for a simple circuit with multiple returns"""
+    dev = qml.device("default.qubit", wires=3)
 
     @qml.batch_params
-    @qml.qnode(dev, diff_method=diff_method)
-    def circuit(x):
+    @qml.qnode(dev)
+    def circuit(data, x, weights):
+        qml.templates.AmplitudeEmbedding(data, wires=[0, 1, 2], normalize=True)
         qml.RX(x, wires=0)
-        qml.RY(0.1, wires=1)
-        qml.CNOT(wires=[0, 1])
-        return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
+        qml.RY(0.2, wires=1)
+        qml.templates.StronglyEntanglingLayers(weights, wires=[0, 1, 2])
+        return qml.expval(qml.PauliZ(0)), qml.probs(wires=[0, 2])
 
-    def cost(x):
-        return np.sum(circuit(x))
-
-    batch_size = 3
+    batch_size = 6
+    data = np.random.random((batch_size, 8))
     x = np.linspace(0.1, 0.5, batch_size, requires_grad=True)
+    weights = np.ones((batch_size, 10, 3, 3), requires_grad=True)
 
-    res = qml.grad(cost)(x)
-    expected = -np.sin(0.1) * np.sin(x)
-    assert np.allclose(res, expected, atol=tol, rtol=0)
+    res = circuit(data, x, weights)
 
+    assert isinstance(res, tuple)
+    assert len(res) == 2
 
-@pytest.mark.jax
-@pytest.mark.parametrize("diff_method", ["backprop", "adjoint", "parameter-shift"])
-@pytest.mark.parametrize("interface", ["jax"])
-def test_jax(diff_method, tol, interface):
-    """Test derivatives when using JAX."""
-    import jax
-
-    jnp = jax.numpy
-    dev = qml.device("default.qubit", wires=2)
-
-    @qml.batch_params
-    @qml.qnode(dev, interface=interface, diff_method=diff_method)
-    def circuit(x):
-        qml.RX(x, wires=0)
-        qml.RY(0.1, wires=1)
-        qml.CNOT(wires=[0, 1])
-        return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-    def cost(x):
-        return jnp.sum(circuit(x))
-
-    batch_size = 3
-    x = jnp.linspace(0.1, 0.5, batch_size)
-
-    res = jax.grad(cost)(x)
-    expected = -np.sin(0.1) * np.sin(x)
-    assert np.allclose(res, expected, atol=tol, rtol=0)
+    assert res[0].shape == (batch_size,)
+    assert res[1].shape == (batch_size, 4)
 
 
-@pytest.mark.jax
-@pytest.mark.parametrize("diff_method", ["adjoint", "parameter-shift"])
-@pytest.mark.parametrize("interface", ["jax-jit"])
-def test_jax_jit(diff_method, interface, tol):
-    """Test derivatives when using JAX and JIT."""
-    import jax
-
-    jnp = jax.numpy
-    dev = qml.device("default.qubit", wires=2)
+def test_shot_vector():
+    """Test that batching works for a simple circuit with a shot vector"""
+    dev = qml.device("default.qubit", wires=3, shots=(100, (200, 3), 300))
 
     @qml.batch_params
-    @qml.qnode(dev, interface=interface, diff_method=diff_method)
-    def circuit(x):
+    @qml.qnode(dev)
+    def circuit(data, x, weights):
+        qml.templates.AmplitudeEmbedding(data, wires=[0, 1, 2], normalize=True)
         qml.RX(x, wires=0)
-        qml.RY(0.1, wires=1)
-        qml.CNOT(wires=[0, 1])
-        return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
+        qml.RY(0.2, wires=1)
+        qml.templates.StronglyEntanglingLayers(weights, wires=[0, 1, 2])
+        return qml.probs(wires=[0, 2])
 
-    @jax.jit
-    def cost(x):
-        return jnp.sum(circuit(x))
+    batch_size = 6
+    data = np.random.random((batch_size, 8))
+    x = np.linspace(0.1, 0.5, batch_size, requires_grad=True)
+    weights = np.ones((batch_size, 10, 3, 3), requires_grad=True)
 
-    batch_size = 3
-    x = jnp.linspace(0.1, 0.5, batch_size)
+    res = circuit(data, x, weights)
 
-    res = jax.grad(cost)(x)
-    expected = -np.sin(0.1) * np.sin(x)
-    assert np.allclose(res, expected, atol=tol, rtol=0)
+    assert isinstance(res, tuple)
+    assert len(res) == 5
+    assert all(shot_res.shape == (batch_size, 4) for shot_res in res)
 
 
-@pytest.mark.torch
-@pytest.mark.parametrize("diff_method", ["backprop", "adjoint", "parameter-shift"])
-@pytest.mark.parametrize("interface", ["torch"])
-def test_torch(diff_method, tol, interface):
-    """Test derivatives when using Torch"""
-    import torch
-
-    dev = qml.device("default.qubit", wires=2)
+def test_multi_returns_shot_vector():
+    """Test that batching works for a simple circuit with multiple returns
+    and with a shot vector"""
+    dev = qml.device("default.qubit", wires=3, shots=(100, (200, 3), 300))
 
     @qml.batch_params
-    @qml.qnode(dev, interface=interface, diff_method=diff_method)
-    def circuit(x):
+    @qml.qnode(dev)
+    def circuit(data, x, weights):
+        qml.templates.AmplitudeEmbedding(data, wires=[0, 1, 2], normalize=True)
         qml.RX(x, wires=0)
-        qml.RY(0.1, wires=1)
-        qml.CNOT(wires=[0, 1])
-        return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
+        qml.RY(0.2, wires=1)
+        qml.templates.StronglyEntanglingLayers(weights, wires=[0, 1, 2])
+        return qml.expval(qml.PauliZ(0)), qml.probs(wires=[0, 2])
 
-    def cost(x):
-        return torch.sum(circuit(x))
+    batch_size = 6
+    data = np.random.random((batch_size, 8))
+    x = np.linspace(0.1, 0.5, batch_size, requires_grad=True)
+    weights = np.ones((batch_size, 10, 3, 3), requires_grad=True)
 
-    batch_size = 3
-    x = torch.linspace(0.1, 0.5, batch_size, requires_grad=True)
+    res = circuit(data, x, weights)
 
-    loss = cost(x)
-    loss.backward()
+    assert isinstance(res, tuple)
+    assert len(res) == 5
+    assert all(isinstance(shot_res, tuple) for shot_res in res)
+    assert all(len(shot_res) == 2 for shot_res in res)
+    assert all(shot_res[0].shape == (batch_size,) for shot_res in res)
+    assert all(shot_res[1].shape == (batch_size, 4) for shot_res in res)
 
-    res = x.grad
-    expected = -np.sin(0.1) * torch.sin(x)
-    assert torch.allclose(res, expected, atol=tol, rtol=0)
 
+class TestDiffSingle:
+    """Test gradients for a single measurement"""
 
-@pytest.mark.tf
-@pytest.mark.parametrize("diff_method", ["backprop", "adjoint", "parameter-shift"])
-@pytest.mark.parametrize("interface", ["tf"])
-def test_tf(diff_method, tol, interface):
-    """Test derivatives when using TF"""
-    import tensorflow as tf
+    @pytest.mark.autograd
+    @pytest.mark.parametrize("diff_method", ["backprop", "adjoint", "parameter-shift"])
+    def test_autograd(self, diff_method, tol):
+        """Test derivatives when using autograd"""
+        dev = qml.device("default.qubit", wires=2)
 
-    dev = qml.device("default.qubit", wires=2)
+        @qml.batch_params
+        @qml.qnode(dev, diff_method=diff_method)
+        def circuit(x):
+            qml.RX(x, wires=0)
+            qml.RY(0.1, wires=1)
+            qml.CNOT(wires=[0, 1])
+            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
 
-    @qml.batch_params
-    @qml.qnode(dev, interface=interface, diff_method=diff_method)
-    def circuit(x):
-        qml.RX(x, wires=0)
-        qml.RY(0.1, wires=1)
-        qml.CNOT(wires=[0, 1])
-        return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
+        def cost(x):
+            return np.sum(circuit(x))
 
-    def cost(x):
-        return tf.reduce_sum(circuit(x))
+        batch_size = 3
+        x = np.linspace(0.1, 0.5, batch_size, requires_grad=True)
 
-    batch_size = 3
-    x = tf.Variable(tf.linspace(0.1, 0.5, batch_size))
+        res = qml.grad(cost)(x)
+        expected = -np.sin(0.1) * np.sin(x)
+        assert np.allclose(res, expected, atol=tol, rtol=0)
 
-    with tf.GradientTape() as tape:
+    @pytest.mark.jax
+    @pytest.mark.parametrize("diff_method", ["backprop", "adjoint", "parameter-shift"])
+    @pytest.mark.parametrize("interface", ["auto", "jax"])
+    def test_jax(self, diff_method, tol, interface):
+        """Test derivatives when using JAX."""
+        import jax
+
+        jnp = jax.numpy
+        dev = qml.device("default.qubit", wires=2)
+
+        @qml.batch_params
+        @qml.qnode(dev, interface=interface, diff_method=diff_method)
+        def circuit(x):
+            qml.RX(x, wires=0)
+            qml.RY(0.1, wires=1)
+            qml.CNOT(wires=[0, 1])
+            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
+
+        def cost(x):
+            return jnp.sum(circuit(x))
+
+        batch_size = 3
+        x = jnp.linspace(0.1, 0.5, batch_size)
+
+        res = jax.grad(cost)(x)
+        expected = -np.sin(0.1) * np.sin(x)
+        assert np.allclose(res, expected, atol=tol, rtol=0)
+
+    @pytest.mark.jax
+    @pytest.mark.parametrize("diff_method", ["adjoint", "parameter-shift"])
+    @pytest.mark.parametrize("interface", ["auto", "jax", "jax-jit"])
+    def test_jax_jit(self, diff_method, interface, tol):
+        """Test derivatives when using JAX and JIT."""
+        import jax
+        import jax.numpy as jnp
+
+        jax.config.update("jax_enable_x64", True)
+
+        dev = qml.device("default.qubit", wires=2)
+
+        @qml.batch_params
+        @qml.qnode(dev, interface=interface, diff_method=diff_method)
+        def circuit(x):
+            qml.RX(x, wires=0)
+            qml.RY(0.1, wires=1)
+            qml.CNOT(wires=[0, 1])
+            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
+
+        @jax.jit
+        def cost(x):
+            return jnp.sum(circuit(x))
+
+        batch_size = 3
+        x = jnp.linspace(0.1, 0.5, batch_size)
+
+        res = jax.grad(cost)(x)
+        expected = -np.sin(0.1) * np.sin(x)
+        assert np.allclose(res, expected, atol=tol, rtol=0)
+
+    @pytest.mark.torch
+    @pytest.mark.parametrize("diff_method", ["backprop", "adjoint", "parameter-shift"])
+    @pytest.mark.parametrize("interface", ["auto", "torch"])
+    def test_torch(self, diff_method, tol, interface):
+        """Test derivatives when using Torch"""
+        import torch
+
+        dev = qml.device("default.qubit", wires=2)
+
+        @qml.batch_params
+        @qml.qnode(dev, interface=interface, diff_method=diff_method)
+        def circuit(x):
+            qml.RX(x, wires=0)
+            qml.RY(0.1, wires=1)
+            qml.CNOT(wires=[0, 1])
+            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
+
+        def cost(x):
+            return torch.sum(circuit(x))
+
+        batch_size = 3
+        x = torch.linspace(0.1, 0.5, batch_size, requires_grad=True)
+
         loss = cost(x)
+        loss.backward()
 
-    res = tape.gradient(loss, x)
-    expected = -np.sin(0.1) * tf.sin(x)
-    assert np.allclose(res, expected, atol=tol, rtol=0)
+        res = x.grad
+        expected = -np.sin(0.1) * torch.sin(x)
+        assert torch.allclose(res, expected, atol=tol, rtol=0)
+
+    @pytest.mark.tf
+    @pytest.mark.parametrize("diff_method", ["backprop", "adjoint", "parameter-shift"])
+    @pytest.mark.parametrize("interface", ["auto", "tf"])
+    def test_tf(self, diff_method, tol, interface):
+        """Test derivatives when using TF"""
+        import tensorflow as tf
+
+        dev = qml.device("default.qubit", wires=2)
+
+        @qml.batch_params
+        @qml.qnode(dev, interface=interface, diff_method=diff_method)
+        def circuit(x):
+            qml.RX(x, wires=0)
+            qml.RY(0.1, wires=1)
+            qml.CNOT(wires=[0, 1])
+            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
+
+        def cost(x):
+            return tf.reduce_sum(circuit(x))
+
+        batch_size = 3
+        x = tf.Variable(tf.linspace(0.1, 0.5, batch_size))
+
+        with tf.GradientTape() as tape:
+            loss = cost(x)
+
+        res = tape.gradient(loss, x)
+        expected = -np.sin(0.1) * tf.sin(x)
+        assert np.allclose(res, expected, atol=tol, rtol=0)
+
+    @pytest.mark.tf
+    @pytest.mark.parametrize("interface", ["auto", "tf", "tf-autograph"])
+    def test_tf_autograph(self, tol, interface):
+        """Test derivatives when using TF and autograph"""
+        import tensorflow as tf
+
+        dev = qml.device("default.qubit", wires=2)
+
+        @qml.batch_params
+        @qml.qnode(dev, interface=interface, diff_method="backprop")
+        def circuit(x):
+            qml.RX(x, wires=0)
+            qml.RY(0.1, wires=1)
+            qml.CNOT(wires=[0, 1])
+            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
+
+        @tf.function
+        def cost(x):
+            return tf.reduce_sum(circuit(x))
+
+        batch_size = 3
+        x = tf.Variable(tf.linspace(0.1, 0.5, batch_size))
+
+        with tf.GradientTape() as tape:
+            loss = cost(x)
+
+        res = tape.gradient(loss, x)
+        expected = -np.sin(0.1) * tf.sin(x)
+        assert np.allclose(res, expected, atol=tol, rtol=0)
 
 
-@pytest.mark.tf
-@pytest.mark.parametrize("interface", ["auto", "tf"])
-def test_tf_autograph(tol, interface):
-    """Test derivatives when using TF and autograph"""
-    import tensorflow as tf
+class TestDiffMulti:
+    """Test gradients for multiple measurements"""
 
-    dev = qml.device("default.qubit", wires=2)
+    @pytest.mark.autograd
+    @pytest.mark.parametrize("diff_method", ["backprop", "parameter-shift"])
+    def test_autograd(self, diff_method, tol):
+        """Test derivatives when using autograd"""
+        dev = qml.device("default.qubit", wires=2)
 
-    @qml.batch_params
-    @qml.qnode(dev, interface=interface, diff_method="backprop")
-    def circuit(x):
-        qml.RX(x, wires=0)
-        qml.RY(0.1, wires=1)
-        qml.CNOT(wires=[0, 1])
-        return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
+        @qml.batch_params
+        @qml.qnode(dev, diff_method=diff_method)
+        def circuit(x):
+            qml.RY(x, wires=0)
+            qml.CNOT(wires=[0, 1])
+            return qml.expval(qml.PauliZ(0)), qml.probs(wires=[0, 1])
 
-    @tf.function
-    def cost(x):
-        return tf.reduce_sum(circuit(x))
+        def cost(x):
+            res = circuit(x)
+            return qml.math.concatenate([qml.math.expand_dims(res[0], 1), res[1]], axis=1)
 
-    batch_size = 3
-    x = tf.Variable(tf.linspace(0.1, 0.5, batch_size))
+        batch_size = 3
+        x = np.linspace(0.1, 0.5, batch_size, requires_grad=True)
 
-    with tf.GradientTape() as tape:
-        loss = cost(x)
+        res = cost(x)
+        expected = qml.math.transpose(
+            qml.math.stack(
+                [
+                    np.cos(x),
+                    np.cos(x / 2) ** 2,
+                    np.zeros_like(x),
+                    np.zeros_like(x),
+                    np.sin(x / 2) ** 2,
+                ]
+            )
+        )
+        assert qml.math.allclose(res, expected, atol=tol)
 
-    res = tape.gradient(loss, x)
-    expected = -np.sin(0.1) * tf.sin(x)
-    assert np.allclose(res, expected, atol=tol, rtol=0)
+        grad = qml.jacobian(cost)(x)
+        expected = qml.math.stack(
+            [-np.sin(x), -np.sin(x) / 2, np.zeros_like(x), np.zeros_like(x), np.sin(x) / 2]
+        ) * qml.math.expand_dims(np.eye(batch_size), 1)
+
+        assert np.allclose(grad, expected, atol=tol, rtol=0)
+
+    @pytest.mark.jax
+    @pytest.mark.parametrize("diff_method", ["backprop", "parameter-shift"])
+    @pytest.mark.parametrize("interface", ["auto", "jax", "jax-jit"])
+    def test_jax(self, diff_method, tol, interface):
+        """Test derivatives when using JAX"""
+        import jax
+        import jax.numpy as jnp
+
+        jax.config.update("jax_enable_x64", True)
+        dev = qml.device("default.qubit", wires=2)
+
+        @qml.batch_params(all_operations=True)
+        @qml.qnode(dev, diff_method=diff_method, interface=interface)
+        def circuit(x):
+            qml.RY(x, wires=0)
+            qml.CNOT(wires=[0, 1])
+            return qml.expval(qml.PauliZ(0)), qml.probs(wires=[0, 1])
+
+        batch_size = 3
+        x = jnp.linspace(0.1, 0.5, batch_size)
+
+        res = circuit(x)
+        expected = (
+            jnp.cos(x),
+            qml.math.transpose(
+                qml.math.stack(
+                    [jnp.cos(x / 2) ** 2, jnp.zeros_like(x), jnp.zeros_like(x), jnp.sin(x / 2) ** 2]
+                )
+            ),
+        )
+
+        assert isinstance(res, tuple)
+        assert len(res) == 2
+        for r, exp in zip(res, expected):
+            assert qml.math.allclose(r, exp, atol=tol)
+
+        grad = jax.jacobian(circuit)(x)
+        expected = (
+            -np.sin(x) * np.eye(batch_size),
+            qml.math.stack([-np.sin(x) / 2, np.zeros_like(x), np.zeros_like(x), np.sin(x) / 2])
+            * qml.math.expand_dims(np.eye(batch_size), 1),
+        )
+
+        assert isinstance(grad, tuple)
+        assert len(grad) == 2
+        for g, exp in zip(grad, expected):
+            assert qml.math.allclose(g, exp, atol=tol, rtol=0)
+
+    @pytest.mark.jax
+    @pytest.mark.parametrize("diff_method", ["parameter-shift"])
+    @pytest.mark.parametrize("interface", ["auto", "jax", "jax-jit"])
+    def test_jax_jit(self, diff_method, tol, interface):
+        """Test derivatives when using JAX"""
+        import jax
+        import jax.numpy as jnp
+
+        jax.config.update("jax_enable_x64", True)
+
+        dev = qml.device("default.qubit", wires=2)
+
+        @jax.jit
+        @qml.batch_params(all_operations=True)
+        @qml.qnode(dev, diff_method=diff_method, interface=interface)
+        def circuit(x):
+            qml.RY(x, wires=0)
+            qml.CNOT(wires=[0, 1])
+            return qml.expval(qml.PauliZ(0)), qml.probs(wires=[0, 1])
+
+        batch_size = 3
+        x = jnp.linspace(0.1, 0.5, batch_size)
+
+        res = circuit(x)
+        expected = (
+            jnp.cos(x),
+            qml.math.transpose(
+                qml.math.stack(
+                    [jnp.cos(x / 2) ** 2, jnp.zeros_like(x), jnp.zeros_like(x), jnp.sin(x / 2) ** 2]
+                )
+            ),
+        )
+
+        assert isinstance(res, tuple)
+        assert len(res) == 2
+        for r, exp in zip(res, expected):
+            assert qml.math.allclose(r, exp, atol=tol)
+
+        grad = jax.jacobian(circuit)(x)
+        expected = (
+            -np.sin(x) * np.eye(batch_size),
+            qml.math.stack([-np.sin(x) / 2, np.zeros_like(x), np.zeros_like(x), np.sin(x) / 2])
+            * qml.math.expand_dims(np.eye(batch_size), 1),
+        )
+
+        assert isinstance(grad, tuple)
+        assert len(grad) == 2
+        for g, exp in zip(grad, expected):
+            assert qml.math.allclose(g, exp, atol=tol, rtol=0)
+
+    @pytest.mark.torch
+    @pytest.mark.parametrize("diff_method", ["backprop", "parameter-shift"])
+    @pytest.mark.parametrize("interface", ["auto", "torch"])
+    def test_torch(self, diff_method, tol, interface):
+        """Test derivatives when using torch"""
+        import torch
+
+        dev = qml.device("default.qubit", wires=2)
+
+        @qml.batch_params
+        @qml.qnode(dev, diff_method=diff_method, interface=interface)
+        def circuit(x):
+            qml.RY(x, wires=0)
+            qml.CNOT(wires=[0, 1])
+            return qml.expval(qml.PauliZ(0)), qml.probs(wires=[0, 1])
+
+        batch_size = 3
+        x = torch.tensor(np.linspace(0.1, 0.5, batch_size), requires_grad=True)
+
+        res = circuit(x)
+        expected = (
+            torch.cos(x),
+            qml.math.transpose(
+                qml.math.stack(
+                    [
+                        torch.cos(x / 2) ** 2,
+                        torch.zeros_like(x),
+                        torch.zeros_like(x),
+                        torch.sin(x / 2) ** 2,
+                    ]
+                )
+            ),
+        )
+
+        assert isinstance(res, tuple)
+        assert len(res) == 2
+        for r, exp in zip(res, expected):
+            assert qml.math.allclose(r, exp, atol=tol)
+
+        grad = torch.autograd.functional.jacobian(circuit, x)
+        expected = (
+            -torch.sin(x) * torch.eye(batch_size),
+            qml.math.stack(
+                [-torch.sin(x) / 2, torch.zeros_like(x), torch.zeros_like(x), torch.sin(x) / 2]
+            )
+            * qml.math.expand_dims(torch.eye(batch_size), 1),
+        )
+
+        assert isinstance(grad, tuple)
+        assert len(grad) == 2
+        for g, exp in zip(grad, expected):
+            assert qml.math.allclose(g, exp, atol=tol, rtol=0)
+
+    @pytest.mark.tf
+    @pytest.mark.parametrize("diff_method", ["backprop", "parameter-shift"])
+    @pytest.mark.parametrize("interface", ["auto", "tf"])
+    def test_tf(self, diff_method, tol, interface):
+        """Test derivatives when using TF"""
+        import tensorflow as tf
+
+        dev = qml.device("default.qubit", wires=2)
+
+        @qml.batch_params
+        @qml.qnode(dev, diff_method=diff_method, interface=interface)
+        def circuit(x):
+            qml.RY(x, wires=0)
+            qml.CNOT(wires=[0, 1])
+            return qml.expval(qml.PauliZ(0)), qml.probs(wires=[0, 1])
+
+        batch_size = 3
+        x = tf.Variable(np.linspace(0.1, 0.5, batch_size))
+
+        with tf.GradientTape() as tape:
+            res = circuit(x)
+            res = qml.math.concatenate([qml.math.expand_dims(res[0], 1), res[1]], axis=1)
+
+        expected = qml.math.transpose(
+            qml.math.stack(
+                [
+                    np.cos(x),
+                    np.cos(x / 2) ** 2,
+                    np.zeros_like(x),
+                    np.zeros_like(x),
+                    np.sin(x / 2) ** 2,
+                ]
+            )
+        )
+        assert qml.math.allclose(res, expected, atol=tol)
+
+        grad = tape.jacobian(res, x)
+        expected = qml.math.stack(
+            [-np.sin(x), -np.sin(x) / 2, np.zeros_like(x), np.zeros_like(x), np.sin(x) / 2]
+        ) * qml.math.expand_dims(np.eye(batch_size), 1)
+
+        assert np.allclose(grad, expected, atol=tol, rtol=0)
+
+    @pytest.mark.tf
+    @pytest.mark.parametrize("diff_method", ["backprop", "parameter-shift"])
+    @pytest.mark.parametrize("interface", ["auto", "tf"])
+    def test_tf_autograph(self, diff_method, tol, interface):
+        """Test derivatives when using TF"""
+        import tensorflow as tf
+
+        dev = qml.device("default.qubit", wires=2)
+
+        @tf.function
+        @qml.batch_params
+        @qml.qnode(dev, diff_method=diff_method, interface=interface)
+        def circuit(x):
+            qml.RY(x, wires=0)
+            qml.CNOT(wires=[0, 1])
+            return qml.expval(qml.PauliZ(0)), qml.probs(wires=[0, 1])
+
+        batch_size = 3
+        x = tf.Variable(np.linspace(0.1, 0.5, batch_size))
+
+        with tf.GradientTape() as tape:
+            res = circuit(x)
+            res = qml.math.concatenate([qml.math.expand_dims(res[0], 1), res[1]], axis=1)
+
+        expected = qml.math.transpose(
+            qml.math.stack(
+                [
+                    np.cos(x),
+                    np.cos(x / 2) ** 2,
+                    np.zeros_like(x),
+                    np.zeros_like(x),
+                    np.sin(x / 2) ** 2,
+                ]
+            )
+        )
+        assert qml.math.allclose(res, expected, atol=tol)
+
+        grad = tape.jacobian(res, x)
+        expected = qml.math.stack(
+            [-np.sin(x), -np.sin(x) / 2, np.zeros_like(x), np.zeros_like(x), np.sin(x) / 2]
+        ) * qml.math.expand_dims(np.eye(batch_size), 1)
+
+        assert np.allclose(grad, expected, atol=tol, rtol=0)
 
 
 def test_all_operations(mocker):
